client_python/requirements.txt
grpcio>=1.75.1
protobuf>=4.25.1